fixes to heatmap
